+++ src/components/button/button.ts
@@ -6,18 +6,38 @@ import {
 } from 'vue';
 import { type Color, type Size } from '../../common';
 
+type Variant = 'solid' | 'outline';
+
 export const classList: {
   base: string;
-  colors: Record<Color, string>;
+  colors: Record<Color, Record<Variant, string>>;
   sizes: Record<Size, string>;
 } = {
   base: 'font-medium cursor-pointer inline-flex items-center justify-center',
   colors: {
-    primary: 'bg-blue-600 text-white hover:bg-blue-700',
-    success: 'bg-green-600 text-white hover:bg-green-700',
-    error: 'bg-red-600 text-white hover:bg-red-700',
-    warning: 'bg-yellow-600 text-white hover:bg-yellow-700',
-    light: 'bg-white text-gray-700 hover:bg-gray-100',
+    primary: {
+      solid: 'bg-blue-600 text-white hover:bg-blue-700',
+      outline:
+        'border-blue-600 text-blue-600 hover:bg-blue-600 hover:text-white',
+    },
+    success: {
+      solid: 'bg-green-600 text-white hover:bg-green-700',
+      outline:
+        'border-green-600 text-green-600 hover:bg-green-600 hover:text-white',
+    },
+    error: {
+      solid: 'bg-red-600 text-white hover:bg-red-700',
+      outline: 'border-red-600 text-red-600 hover:bg-red-600 hover:text-white',
+    },
+    warning: {
+      solid: 'bg-yellow-600 text-white hover:bg-yellow-700',
+      outline:
+        'border-yellow-600 text-yellow-600 hover:bg-yellow-600 hover:text-white',
+    },
+    light: {
+      solid: 'bg-white text-gray-700 hover:bg-gray-100',
+      outline: 'border-gray-300 text-gray-900 hover:bg-gray-100',
+    },
   },
   sizes: {
     sm: 'h-8 px-3 rounded text-sm',
@@ -30,6 +50,7 @@ type ButtonProps = {
   color?: Color;
   size?: Size;
   tag?: string | Component;
+  variant?: Variant;
 };
 
 const Button: FunctionalComponent<ButtonProps> = (props, context) => {
@@ -40,8 +61,9 @@ const Button: FunctionalComponent<ButtonProps> = (props, context) => {
       class: [
         inheritClass,
         classList.base,
-        classList.colors[props.color ?? 'primary'],
+        classList.colors[props.color ?? 'primary'][props.variant ?? 'solid'],
         classList.sizes[props.size ?? 'md'],
+        props.variant === 'solid' ? '' : 'border',
       ],
       ...inheritAttributes,
     },
@@ -63,6 +85,10 @@ Button.props = {
   tag: {
     type: null,
     default: 'button',
+  },
+  variant: {
+    type: String as PropType<ButtonProps['variant']>,
+    default: 'solid',
   },
 };