+++ src/main.ts
@@ -3,5 +3,25 @@ import Input from './components/input/input';
 import Checkbox from './components/checkbox/checkbox';
 import Radio from './components/radio/radio';
 import Select from './components/select/select';
+import FormItem from './components/form-item/form-item';
+import Modal from './components/modal/modal';
+import Table from './components/table/table';
+import Widget from './components/widget/widget';
+import Pagination from './components/pagination/pagination';
+import Message from './components/message/message';
+import { useMessage } from './components/message/message.compose';
 
-export { Button, Input, Checkbox, Radio, Select };
+export {
+  Button,
+  Input,
+  Checkbox,
+  Radio,
+  Select,
+  FormItem,
+  Modal,
+  Table,
+  Widget,
+  Pagination,
+  Message,
+  useMessage,
+};