+++ CHANGELOG.md
@@ -7,18 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-06-06
+
 ### Added
 
-- Form item component
-- Modal component
-- Table component
-- Widget component
+- Message component (#29).
+- Pagination component (#28).
+- Input currency variant (#27).
+- Modal size variant (#26).
+- Custom table empty message (#23).
+
+### Fixed
+
+- Select option null not selected (#24).
+
+## [0.2.0] - 2026-05-25
+
+### Added
+
+- Widget component (#21).
+- Table component (#20).
+- Modal component (#19).
+- Form item component (#18).
+- Input Textarea (#17).
 
 ### Fixed
 
 - Add outline focus to input and select (#16).
 
-## v0.1.3 - 2026-05-22
+## [0.1.3] - 2026-05-22
 
 ### Fixed